further refactor the message structs
