+++ rust/sglang-server/src/message/egress.rs
@@ -1,10 +1,59 @@
-//! The egress-ring wire: frame encodings (batch / control result / error), the
-//! columnar batch decode into per-request [`ChunkEvent`]s, and the chunk types.
+//! The egress (response) direction: the per-request back-channel the API
+//! handler drains ([`EgressSink`] / [`EgressItem`]), the egress-ring frame
+//! encodings (batch / control result / error), and the columnar batch decode
+//! into per-request [`ChunkEvent`]s.
 
 use bytes::Bytes;
 use serde::{Deserialize, Serialize};
+use tokio::sync::mpsc;
 
+use crate::error::Error;
 use crate::ids::RidHash;
+
+/// Per-request back-channel the detok shard writes egress frames to and the API
+/// handler drains for SSE; bounded, and receiver-drop (disconnect) = stream end.
+#[derive(Clone, Debug)]
+pub enum EgressSink {
+    Local(mpsc::Sender<EgressItem>),
+}
+
+/// Why an [`EgressSink::try_send`] failed: `Full` = client backpressure, `Closed`
+/// = client gone. Both terminal for a stream; the caller distinguishes for logging.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum SinkError {
+    Full,
+    Closed,
+}
+
+impl EgressSink {
+    /// Non-blocking send. `Err(Full)` = backpressure, `Err(Closed)` = client gone.
+    pub fn try_send(&self, item: EgressItem) -> Result<(), SinkError> {
+        match self {
+            EgressSink::Local(tx) => tx.try_send(item).map_err(|e| match e {
+                mpsc::error::TrySendError::Full(_) => SinkError::Full,
+                mpsc::error::TrySendError::Closed(_) => SinkError::Closed,
+            }),
+        }
+    }
+}
+
+#[allow(dead_code)] // the receiver half is created inline in api_server::submit.
+pub type EgressSource = mpsc::Receiver<EgressItem>;
+
+/// What the connection handler receives on the egress stream: a detok-decoded
+/// [`ChunkEvent`] (handler formats it), a verbatim control payload, or an error.
+#[derive(Debug)]
+pub enum EgressItem {
+    /// An intermediate streamed generation step (only sent for streaming reqs).
+    Frame(ChunkEvent),
+    /// The final generation step.
+    Done(ChunkEvent),
+    /// A control-request result: one verbatim payload (e.g. `/server_info`),
+    /// delivered as-is with no per-protocol formatting.
+    Control(Bytes),
+    /// Terminal failure: handler emits an error frame (stream) or status (unary).
+    Error(Error),
+}
 
 /// Egress-ring frame tag (first byte, prepended Rust-side; Python wire unchanged):
 /// a single control-request result payload.
